fix login, join
